Move cardNumber StateIIN helpers to StateIIN.kt
These helpers are going to be consumed by more than just the
ElementStateManagers so putting them in a more centrally located home
to prevent random modules importing from ElementStateManager's module
would likely improve code organization
