+++ forage-android/src/main/java/com/joinforage/forage/android/core/element/state/PanElementStateManager.kt
@@ -4,32 +4,15 @@ import com.joinforage.forage.android.core.element.ElementValidationError
 import com.joinforage.forage.android.core.element.IncompleteEbtPanError
 import com.joinforage.forage.android.core.element.InvalidEbtPanError
 import com.joinforage.forage.android.core.element.TooLongEbtPanError
-import com.joinforage.forage.android.model.STATE_INN_LENGTH
-import com.joinforage.forage.android.model.StateIIN
+import com.joinforage.forage.android.model.hasInvalidStateIIN
+import com.joinforage.forage.android.model.isCorrectLength
+import com.joinforage.forage.android.model.missingStateIIN
+import com.joinforage.forage.android.model.tooLongForStateIIN
+import com.joinforage.forage.android.model.tooShortForStateIIN
 
 const val MIN_CARD_LENGTH = 16
 const val MAX_CARD_LENGTH = 19
 
-private fun missingStateIIN(cardNumber: String): Boolean {
-    return cardNumber.length < STATE_INN_LENGTH
-}
-private fun queryForStateIIN(cardNumber: String): StateIIN? {
-    return StateIIN.values().find { cardNumber.startsWith(it.iin) }
-}
-private fun hasInvalidStateIIN(cardNumber: String): Boolean {
-    return queryForStateIIN(cardNumber) == null
-}
-private fun tooShortForStateIIN(cardNumber: String): Boolean {
-    val iin = queryForStateIIN(cardNumber) ?: return false
-    return cardNumber.length < iin.panLength
-}
-private fun tooLongForStateIIN(cardNumber: String): Boolean {
-    val iin = queryForStateIIN(cardNumber) ?: return false
-    return cardNumber.length > iin.panLength
-}
-private fun isCorrectLength(cardNumber: String): Boolean {
-    return !tooShortForStateIIN(cardNumber) && !tooLongForStateIIN(cardNumber)
-}
 private fun failsValidation(cardNumber: String): Boolean {
     return missingStateIIN(cardNumber) ||
         hasInvalidStateIIN(cardNumber) ||
